Move interface types into their own header
