+++ frontend/include/chpl/resolution/resolution-types.h
@@ -2876,111 +2876,6 @@ class ResolvedFields {
   }
 };
 
-/* A pair of (implemented interface, 'implements' statement) IDs derived
-   from a single occurrences of an interface in an inheritance expression list. */
-class ImplementationPoint {
- private:
-  // The interface being implemented (instantiated with the types of the arguments)
-  const types::InterfaceType* interface_;
-  // The ID of the implementation statement
-  ID id_;
-
-  ImplementationPoint(const types::InterfaceType* interface,
-                      ID id)
-    : interface_(interface), id_(id) {}
-
-  static owned<ImplementationPoint> const&
-  getImplementationPoint(Context* context,
-                         const types::InterfaceType* interface,
-                         ID id);
-
- public:
-  static const ImplementationPoint*
-  get(Context* context, const types::InterfaceType* interface, ID id);
-
-  static bool update(owned<ImplementationPoint>& lhs,
-                     owned<ImplementationPoint>& rhs) {
-    return defaultUpdateOwned(lhs, rhs);
-  }
-  bool operator==(const ImplementationPoint& other) const {
-    return interface_ == other.interface_ &&
-           id_ == other.id_;
-  }
-  bool operator!=(const ImplementationPoint& other) const {
-    return !(*this == other);
-  }
-  void mark(Context* context) const;
-  void stringify(std::ostream& ss, chpl::StringifyKind stringKind) const;
-
-  const types::InterfaceType* interface() const { return interface_; }
-
-  const ID& id() const { return id_; }
-};
-
-class ImplementationWitness {
- public:
-  using ConstraintMap = std::unordered_map<ID, const ImplementationWitness*>;
-  using AssociatedTypeMap = types::PlaceholderMap;
-  using FunctionMap = std::unordered_map<ID, ID>;
-
- private:
-
-  ConstraintMap associatedConstraints_;
-  AssociatedTypeMap associatedTypes_;
-  FunctionMap requiredFns_;
-
-  ImplementationWitness(ConstraintMap associatedConstraints,
-                        AssociatedTypeMap associatedTypes,
-                        FunctionMap requiredFns)
-    : associatedConstraints_(std::move(associatedConstraints)),
-      associatedTypes_(std::move(associatedTypes)),
-      requiredFns_(std::move(requiredFns)) {}
-
-  static const owned<ImplementationWitness>&
-  getImplementationWitness(Context* context,
-                           ConstraintMap associatedConstraints,
-                           AssociatedTypeMap associatedTypes,
-                           FunctionMap requiredFns);
-
- public:
-  static ImplementationWitness*
-  get(Context* context,
-      ConstraintMap associatedConstraints,
-      AssociatedTypeMap associatedTypes,
-      FunctionMap requiredFns);
-
-  static bool update(owned<ImplementationWitness>& lhs,
-                     owned<ImplementationWitness>& rhs) {
-    return defaultUpdateOwned(lhs, rhs);
-  }
-  bool operator==(const ImplementationWitness& other) const {
-    return associatedConstraints_ == other.associatedConstraints_ &&
-           associatedTypes_ == other.associatedTypes_ &&
-           requiredFns_ == other.requiredFns_;
-  }
-  bool operator!=(const ImplementationWitness& other) const {
-    return !(*this == other);
-  }
-  void mark(Context* context) const {
-    chpl::mark<decltype(associatedTypes_)>{}(context, associatedTypes_);
-    chpl::mark<decltype(requiredFns_)>{}(context, requiredFns_);
-    chpl::mark<decltype(associatedConstraints_)>{}(context, associatedConstraints_);
-  }
-  void stringify(std::ostream& ss, chpl::StringifyKind stringKind) const;
-
-  const ConstraintMap& associatedConstraints() const {
-    return associatedConstraints_;
-  }
-
-  const AssociatedTypeMap& associatedTypes() const {
-    return associatedTypes_;
-  }
-
-  const FunctionMap& requiredFns() const {
-    return requiredFns_;
-  }
-};
-
 class ResolvedParamLoop {
   public:
     using LoopBodies = std::vector<ResolutionResultByPostorderID>;
@@ -3307,8 +3202,6 @@ CHPL_DEFINE_STD_HASH_(MostSpecificCandidate, (key.hash()));
 CHPL_DEFINE_STD_HASH_(MostSpecificCandidates, (key.hash()));
 CHPL_DEFINE_STD_HASH_(CallResolutionResult, (key.hash()));
 CHPL_DEFINE_STD_HASH_(TheseResolutionResult, (key.hash()));
-CHPL_DEFINE_STD_HASH_(ImplementationWitness::FunctionMap, (chpl::hashUnorderedMap(key)));
-CHPL_DEFINE_STD_HASH_(ImplementationWitness::ConstraintMap, (chpl::hashUnorderedMap(key)));
 #undef CHPL_DEFINE_STD_HASH_
 
 } // end namespace std
