+++ frontend/lib/resolution/resolution-types.cpp
@@ -906,82 +906,6 @@ void ResolvedFields::finalizeFields(Context* context, bool syntaxOnly) {
   allGenericFieldsHaveDefaultValues_ = allGenHaveDefault;
 }
 
-owned<ImplementationPoint> const&
-ImplementationPoint::getImplementationPoint(Context* context, const InterfaceType* interface,
-                                            ID id) {
-  QUERY_BEGIN(getImplementationPoint, context, interface, id);
-  auto result = toOwned(new ImplementationPoint(interface, std::move(id)));
-  return QUERY_END(result);
-}
-
-const ImplementationPoint*
-ImplementationPoint::get(Context* context, const InterfaceType* interface,
-                         ID id) {
-  return getImplementationPoint(context, interface, std::move(id)).get();
-}
-
-void ImplementationPoint::mark(Context* context) const {
-  interface_->mark(context);
-  id_.mark(context);
-}
-
-void ImplementationPoint::stringify(std::ostream& ss, chpl::StringifyKind stringKind) const {
-  ss << "implements(";
-  interface_->stringify(ss, stringKind);
-  ss << " via )";
-  id_.stringify(ss, stringKind);
-}
-
-const owned<ImplementationWitness>&
-ImplementationWitness::getImplementationWitness(Context* context,
-                                                ConstraintMap associatedConstraints,
-                                                AssociatedTypeMap associatedTypes,
-                                                FunctionMap requiredFns) {
-  QUERY_BEGIN(getImplementationWitness, context, associatedConstraints,
-              associatedTypes, requiredFns);
-
-  auto result = toOwned(new ImplementationWitness(std::move(associatedConstraints),
-                                                  std::move(associatedTypes),
-                                                  std::move(requiredFns)));
-
-  return QUERY_END(result);
-}
-
-ImplementationWitness* ImplementationWitness::get(Context* context,
-                                                  ConstraintMap associatedConstraints,
-                                                  AssociatedTypeMap associatedTypes,
-                                                  FunctionMap requiredFns) {
-  return getImplementationWitness(context, std::move(associatedConstraints),
-                                  std::move(associatedTypes),
-                                  std::move(requiredFns)).get();
-}
-
-void ImplementationWitness::stringify(std::ostream& ss, chpl::StringifyKind stringKind) const {
-  ss << "witness(";
-  ss << "constraints: ";
-  for (auto& c : associatedConstraints_) {
-    c.first.stringify(ss, stringKind);
-    ss << " => ";
-    c.second->stringify(ss, stringKind);
-    ss << ", ";
-  }
-  ss << "associated types: ";
-  for (auto& a : associatedTypes_) {
-    a.first.stringify(ss, stringKind);
-    ss << " => ";
-    a.second->stringify(ss, stringKind);
-    ss << ", ";
-  }
-  ss << "required functions: ";
-  for (auto& f : requiredFns_) {
-    f.first.stringify(ss, stringKind);
-    ss << " => ";
-    f.second.stringify(ss, stringKind);
-    ss << ", ";
-  }
-  ss << ")";
-}
-
 const owned<TypedFnSignature>&
 TypedFnSignature::getTypedFnSignature(Context* context,
                     const UntypedFnSignature* untypedSignature,
